requirements.txt
fastapi==0.85
uvicorn[standard]==0.22.0
pydantic[email]==1.10.7
deta~=1.1.0
python-multipart
Pillow==9.5.0
openpyxl
pdf2image